+++ examples/list-resources/tfcoremock_complex_resource/list-resource.tf
@@ -1,0 +1,6 @@
+list "tfcoremock_complex_resource" "example" {
+  provider = tfcoremock
+  config {
+    id = "" # optional
+  }
+}