+++ internal/provider/testdata/list/simple/main.tf
@@ -1,0 +1,10 @@
+
+provider "tfcoremock" {}
+
+resource "tfcoremock_simple_resource" "one" {
+  id = "one"
+}
+
+resource "tfcoremock_simple_resource" "two" {
+  id = "two"
+}